requirements.txt
influxdb-client3
python-dotenv
